requirements.txt
Django==3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
gunicorn==20.1.0
psycopg2-binary==2.9.1
celery==5.1.2
redis==3.5.3
backports.zoneinfo==0.2.1
